requirements.txt
and==66.0.2
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-heroku==0.3.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
gunicorn==20.1.0
phonenumbers==8.13.6
Pillow==9.3.0
pipenv==2022.11.11
platformdirs==2.5.2
pluggy==1.0.0
protobuf==4.21.9
psycopg2==2.9.5
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
pygame==2.1.2
pyhcl==0.4.4
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
pyperclip==1.8.2
pytest==7.1.3
python-crontab==2.4.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
redis==3.4.1
requests==2.28.1
s3transfer==0.6.0
scikit-learn==1.1.3
scipy==1.9.3
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.2
submit50==3.1.1
termcolor==1.1.0
tf==1.0.0
threadpoolctl==3.1.0
tomli==2.0.1
urllib3==1.26.11
validate-email==1.3
validators==0.20.0
vine==1.3.0
virtualenv==20.16.6
virtualenv-clone==0.5.7
whitenoise==5.2.0
zipp==2.2.0
